requirements.txt
pillow
requests
urllib3==1.26.6 